requirements.txt
replicate
pillow
numpy
torch
google-genai>=1.0.0
opencv-python
torchaudio